requirements.txt
py-cord
rich
requests
prettytable
rcat
python-dotenv
pytz
